upgrade to CLDR::Number v0.16
because it—yet again—breaks this project’s unit tests due to the new
“minimum grouping digits” feature (introduced in CLDR v26) finally being
implemented, which causes some locales to not use the grouping symbol until a
specified minimum number of digits.

for example, english and german are 1,000 and 1.000, respectively, but spanish
(excluding latin american spanish) has changed from 1.000 to 1000 and the
grouping symbol isn’t included until one more digit is added: 10.000.